+++ commands/delete_test.go
@@ -41,7 +41,7 @@ func TestDelete(t *testing.T) {
 	if actualRet != expectedRet {
 		t.Fatalf("Main() = %q, want %q", actualRet, expectedRet)
 	}
-	results, err := readPasswords(db, "", "", "", "", false, false, []string{})
+	results, err := readPasswords(db, searchOptions{})
 	if err != nil {
 		t.Fatalf("readPasswords() err = %q, want nil", err)
 	}
@@ -93,7 +93,7 @@ func TestInteractiveDelete(t *testing.T) {
 	if outBuf.String() != expectedBuf {
 		t.Fatalf("Main() output is %q, want %q", outBuf.String(), expectedBuf)
 	}
-	results, err := readPasswords(db, "", "", "", "", false, false, []string{})
+	results, err := readPasswords(db, searchOptions{})
 	if err != nil {
 		t.Fatalf("readPasswords() err = %q, want nil", err)
 	}