+++ guide/src/news.md
@@ -3,6 +3,7 @@
 ## main
 
 - when specifying TOTP shared secrets, it is now supported to specify `otpauth://` URLs
+- new `cpm --qrcode` option to show the TOTP shared secret as a qr code
 
 ## 7.0